// src/app/shared/enum/button-type.enum.ts
export const enum ButtonTypeEnum {
  add = 'add',
  save = 'save',
  cancel = 'cancel',
  search = 'search',
  searchSubmit = 'searchSubmit',
  menu = 'menu',
  send = 'send',
  edit = 'edit',
  preview = 'preview',
  accept = 'accept',
  reject = 'reject',
  sendInvite = 'sendInvite',
  terminate = 'terminate',
  uninvite = 'uninvite',
  // Don't forget to define your button in PageHeaderComponent
}